roles y permisos terminado

+++ app/Http/Controllers/PostController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Post;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
@@ -14,6 +15,7 @@ class PostController extends Controller
      */
     public function index()
     {
+        abort_if(Gate::denies('post_index'), 403);
         $posts = Post::paginate(5);
         return view('posts.index', compact('posts'));
     }
@@ -25,6 +27,7 @@ class PostController extends Controller
      */
     public function create()
     {
+        abort_if(Gate::denies('post_create'), 403);
         return view('posts.create');
     }
 
@@ -48,6 +51,7 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        abort_if(Gate::denies('post_show'), 403);
         return view('posts.show', compact('post'));
     }
 
@@ -58,6 +62,7 @@ class PostController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function edit(Post $post){
+        abort_if(Gate::denies('post_edit'), 403);
         return view('posts.edit', compact('post'));
     }
 
@@ -82,6 +87,7 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        abort_if(Gate::denies('post_destroy'), 403);
         $post->delete();
         return redirect()->route('posts.index');
     }
