+++ app/Http/Controllers/UserController.php
@@ -2,20 +2,26 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\UserCreateRequest;
-use App\Http\Requests\UserEditRequest;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Spatie\Permission\Models\Role;
+use Illuminate\Support\Facades\Gate;
+use App\Http\Requests\UserEditRequest;
+use App\Http\Requests\UserCreateRequest;
 
 class UserController extends Controller
 {
     public function index(){
+        abort_if(Gate::denies('user_index'), 403);
         $users = User::paginate(5);
         return view('users.index', compact('users'));
     }
 
-    public function create(){
-        return view('users.create');
+    public function create()
+    {
+        abort_if(Gate::denies('user_create'), 403);
+        $roles = Role::all()->pluck('name', 'id');
+        return view('users.create', compact('roles'));
     }
 
     public function store(UserCreateRequest $request){
@@ -29,16 +35,24 @@ class UserController extends Controller
         +[
             'password'=>bcrypt($request->input('password'))
         ]);
-            return redirect()->route('users.show', $user->id)->with('success', 'Usuario creado correctamente');
+
+        $roles =$request->input('roles', []);
+        $user->roles()->sync($roles);
+        return redirect()->route('users.show', $user->id)->with('success', 'Usuario creado correctamente');
     }
 
     public function show(User $user){
+        abort_if(Gate::denies('user_show'), 403);
         // $user = User::findOrFail($id);
+        $user->load('roles');
         return view('users.show', compact('user'));
     }
 
     public function edit(User $user){
-        return view('users.edit', compact('user'));
+        abort_if(Gate::denies('user_edit'), 403);
+        $roles = Role::all()->pluck('name', 'id');
+        $user->load('roles');
+        return view('users.edit', compact('user', 'roles'));
     }
 
     public function update(UserEditRequest $request, User $user){
@@ -57,10 +71,16 @@ class UserController extends Controller
         // }
 
         $user->update($data);
+        $roles = $request->input('roles', []);
+        $user->roles()->sync($roles);
         return redirect()->route('users.show', $user->id)->with('success', 'Usuario actualizado correctamente');
     }
 
     public function destroy(User $user){
+        abort_if(Gate::denies('user_destroy'), 403);
+        if(auth()->user()->id = $user->id){
+            return redirect()->route('users.index');
+        }
         $user->delete();
         return back()->with('success', 'Usuario eliminado correctamente');
     }
